pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "skynet-simulation"
version = "0.1.0"
description = "CARLA Simulator Integration for Autonomous Driving System"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "See LICENSE file" }
authors = [{ name = "Skynet" }]
keywords = [
    "autonomous-driving",
    "carla",
    "simulation",
    "vehicle-control",
]

# Core dependencies
dependencies = [
    # Numerical Computing
    "numpy>=1.24.0,<3.0.0",

    # Computer Vision (for visualization)
    "opencv-python>=4.8.0,<5.0.0",

    # CARLA Simulator Python API
    "carla==0.9.16",

    # Image Processing
    "pillow>=10.0.0,<13.0.0",

    # Configuration Management
    "pyyaml>=6.0,<7.0",

    # Distributed System Dependencies
    "pyzmq>=25.1.0,<28.0.0",

    # Visualization
    "pygame>=2.5.0,<3.0.0",

    # Inter-module dependencies
    "skynet-lkas @ git+https://github.com/ADS-Skynet/Lkas.git@split-module",
    "skynet-common @ git+https://github.com/ADS-Skynet/Common.git",
]

# Optional dependency groups
[project.optional-dependencies]
# Development tools
dev = [
    "pytest>=7.4.0,<8.0.0",
    "black>=23.0.0,<25.0.0",
    "flake8>=6.0.0,<8.0.0",
    "mypy>=1.4.0,<2.0.0",
    "isort>=5.12.0,<6.0.0",
]

# All optional dependencies
all = ["skynet-simulation[dev]"]

# Entry points for command-line scripts
[project.scripts]
simulation = "simulation.run:main"

# Package discovery (src-layout)
[tool.setuptools]
package-dir = {"simulation" = "src"}
packages = ["simulation", "simulation.integration", "simulation.processing", "simulation.utils"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml", "*.json"]

# Black code formatter configuration
[tool.black]
line-length = 100
target-version = ['py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# isort import sorter configuration
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

# pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = ["-v", "--strict-markers", "--tb=short"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "carla: marks tests that require CARLA server running",
]

# mypy type checker configuration
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true

[[tool.mypy.overrides]]
module = ["carla.*", "cv2.*", "pygame.*"]
ignore_missing_imports = true
